Migrate internal/proto6 Schema and Attribute functions to unexported tfsdk receiver methods
