+++ tfsdk/attribute_test.go
@@ -1,463 +1,22 @@
-package proto6
+package tfsdk
 
 import (
 	"context"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-framework/attr"
-	"github.com/hashicorp/terraform-plugin-framework/schema"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-
 	"github.com/google/go-cmp/cmp"
+	"github.com/hashicorp/terraform-plugin-framework/attr"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 )
 
-func TestSchema(t *testing.T) {
-	t.Parallel()
-
-	type testCase struct {
-		input       schema.Schema
-		expected    *tfprotov6.Schema
-		expectedErr string
-	}
-
-	tests := map[string]testCase{
-		"empty-val": {
-			input:       schema.Schema{},
-			expectedErr: "must have at least one attribute in the schema",
-		},
-		"basic-attrs": {
-			input: schema.Schema{
-				Version: 1,
-				Attributes: map[string]schema.Attribute{
-					"string": {
-						Type:     types.StringType,
-						Required: true,
-					},
-					"number": {
-						Type:     types.NumberType,
-						Optional: true,
-					},
-					"bool": {
-						Type:     types.BoolType,
-						Computed: true,
-					},
-				},
-			},
-			expected: &tfprotov6.Schema{
-				Version: 1,
-				Block: &tfprotov6.SchemaBlock{
-					Attributes: []*tfprotov6.SchemaAttribute{
-						{
-							Name:     "bool",
-							Type:     tftypes.Bool,
-							Computed: true,
-						},
-						{
-							Name:     "number",
-							Type:     tftypes.Number,
-							Optional: true,
-						},
-						{
-							Name:     "string",
-							Type:     tftypes.String,
-							Required: true,
-						},
-					},
-				},
-			},
-		},
-		"complex-attrs": {
-			input: schema.Schema{
-				Version: 2,
-				Attributes: map[string]schema.Attribute{
-					"list": {
-						Type:     types.ListType{ElemType: types.StringType},
-						Required: true,
-					},
-					"object": {
-						Type: types.ObjectType{AttrTypes: map[string]attr.Type{
-							"string": types.StringType,
-							"number": types.NumberType,
-							"bool":   types.BoolType,
-						}},
-						Optional: true,
-					},
-					"map": {
-						Type:     types.MapType{ElemType: types.NumberType},
-						Computed: true,
-					},
-					// TODO: add tuple support when it lands
-					// TODO: add set support when it lands
-				},
-			},
-			expected: &tfprotov6.Schema{
-				Version: 2,
-				Block: &tfprotov6.SchemaBlock{
-					Attributes: []*tfprotov6.SchemaAttribute{
-						{
-							Name:     "list",
-							Type:     tftypes.List{ElementType: tftypes.String},
-							Required: true,
-						},
-						{
-							Name:     "map",
-							Type:     tftypes.Map{AttributeType: tftypes.Number},
-							Computed: true,
-						},
-						{
-							Name: "object",
-							Type: tftypes.Object{AttributeTypes: map[string]tftypes.Type{
-								"string": tftypes.String,
-								"number": tftypes.Number,
-								"bool":   tftypes.Bool,
-							}},
-							Optional: true,
-						},
-					},
-				},
-			},
-		},
-		"nested-attrs": {
-			input: schema.Schema{
-				Version: 3,
-				Attributes: map[string]schema.Attribute{
-					"single": {
-						Attributes: schema.SingleNestedAttributes(map[string]schema.Attribute{
-							"string": {
-								Type:     types.StringType,
-								Required: true,
-							},
-							"number": {
-								Type:     types.NumberType,
-								Optional: true,
-							},
-							"bool": {
-								Type:     types.BoolType,
-								Computed: true,
-							},
-							"list": {
-								Type:     types.ListType{ElemType: types.StringType},
-								Computed: true,
-								Optional: true,
-							},
-						}),
-						Required: true,
-					},
-					"list": {
-						Attributes: schema.ListNestedAttributes(map[string]schema.Attribute{
-							"string": {
-								Type:     types.StringType,
-								Required: true,
-							},
-							"number": {
-								Type:     types.NumberType,
-								Optional: true,
-							},
-							"bool": {
-								Type:     types.BoolType,
-								Computed: true,
-							},
-							"list": {
-								Type:     types.ListType{ElemType: types.StringType},
-								Computed: true,
-								Optional: true,
-							},
-						}, schema.ListNestedAttributesOptions{}),
-						Optional: true,
-					},
-					"set": {
-						Attributes: schema.SetNestedAttributes(map[string]schema.Attribute{
-							"string": {
-								Type:     types.StringType,
-								Required: true,
-							},
-							"number": {
-								Type:     types.NumberType,
-								Optional: true,
-							},
-							"bool": {
-								Type:     types.BoolType,
-								Computed: true,
-							},
-							"list": {
-								Type:     types.ListType{ElemType: types.StringType},
-								Computed: true,
-								Optional: true,
-							},
-						}, schema.SetNestedAttributesOptions{}),
-						Computed: true,
-					},
-					"map": {
-						Attributes: schema.MapNestedAttributes(map[string]schema.Attribute{
-							"string": {
-								Type:     types.StringType,
-								Required: true,
-							},
-							"number": {
-								Type:     types.NumberType,
-								Optional: true,
-							},
-							"bool": {
-								Type:     types.BoolType,
-								Computed: true,
-							},
-							"list": {
-								Type:     types.ListType{ElemType: types.StringType},
-								Computed: true,
-								Optional: true,
-							},
-						}, schema.MapNestedAttributesOptions{}),
-						Optional: true,
-						Computed: true,
-					},
-				},
-			},
-			expected: &tfprotov6.Schema{
-				Version: 3,
-				Block: &tfprotov6.SchemaBlock{
-					Attributes: []*tfprotov6.SchemaAttribute{
-						{
-							Name: "list",
-							NestedType: &tfprotov6.SchemaObject{
-								Nesting: tfprotov6.SchemaObjectNestingModeList,
-								Attributes: []*tfprotov6.SchemaAttribute{
-									{
-										Name:     "bool",
-										Type:     tftypes.Bool,
-										Computed: true,
-									},
-									{
-										Name:     "list",
-										Type:     tftypes.List{ElementType: tftypes.String},
-										Optional: true,
-										Computed: true,
-									},
-									{
-										Name:     "number",
-										Type:     tftypes.Number,
-										Optional: true,
-									},
-									{
-										Name:     "string",
-										Type:     tftypes.String,
-										Required: true,
-									},
-								},
-							},
-							Optional: true,
-						},
-						{
-							Name: "map",
-							NestedType: &tfprotov6.SchemaObject{
-								Nesting: tfprotov6.SchemaObjectNestingModeMap,
-								Attributes: []*tfprotov6.SchemaAttribute{
-									{
-										Name:     "bool",
-										Type:     tftypes.Bool,
-										Computed: true,
-									},
-									{
-										Name:     "list",
-										Type:     tftypes.List{ElementType: tftypes.String},
-										Optional: true,
-										Computed: true,
-									},
-									{
-										Name:     "number",
-										Type:     tftypes.Number,
-										Optional: true,
-									},
-									{
-										Name:     "string",
-										Type:     tftypes.String,
-										Required: true,
-									},
-								},
-							},
-							Optional: true,
-							Computed: true,
-						},
-						{
-							Name: "set",
-							NestedType: &tfprotov6.SchemaObject{
-								Nesting: tfprotov6.SchemaObjectNestingModeSet,
-								Attributes: []*tfprotov6.SchemaAttribute{
-									{
-										Name:     "bool",
-										Type:     tftypes.Bool,
-										Computed: true,
-									},
-									{
-										Name:     "list",
-										Type:     tftypes.List{ElementType: tftypes.String},
-										Optional: true,
-										Computed: true,
-									},
-									{
-										Name:     "number",
-										Type:     tftypes.Number,
-										Optional: true,
-									},
-									{
-										Name:     "string",
-										Type:     tftypes.String,
-										Required: true,
-									},
-								},
-							},
-							Computed: true,
-						},
-						{
-							Name: "single",
-							NestedType: &tfprotov6.SchemaObject{
-								Nesting: tfprotov6.SchemaObjectNestingModeSingle,
-								Attributes: []*tfprotov6.SchemaAttribute{
-									{
-										Name:     "bool",
-										Type:     tftypes.Bool,
-										Computed: true,
-									},
-									{
-										Name:     "list",
-										Type:     tftypes.List{ElementType: tftypes.String},
-										Optional: true,
-										Computed: true,
-									},
-									{
-										Name:     "number",
-										Type:     tftypes.Number,
-										Optional: true,
-									},
-									{
-										Name:     "string",
-										Type:     tftypes.String,
-										Required: true,
-									},
-								},
-							},
-							Required: true,
-						},
-					},
-				},
-			},
-		},
-		"markdown-description": {
-			input: schema.Schema{
-				Version: 1,
-				Attributes: map[string]schema.Attribute{
-					"string": {
-						Type:     types.StringType,
-						Required: true,
-					},
-				},
-				MarkdownDescription: "a test resource",
-			},
-			expected: &tfprotov6.Schema{
-				Version: 1,
-				Block: &tfprotov6.SchemaBlock{
-					Attributes: []*tfprotov6.SchemaAttribute{
-						{
-							Name:     "string",
-							Type:     tftypes.String,
-							Required: true,
-						},
-					},
-					Description:     "a test resource",
-					DescriptionKind: tfprotov6.StringKindMarkdown,
-				},
-			},
-		},
-		"plaintext-description": {
-			input: schema.Schema{
-				Version: 1,
-				Attributes: map[string]schema.Attribute{
-					"string": {
-						Type:     types.StringType,
-						Required: true,
-					},
-				},
-				Description: "a test resource",
-			},
-			expected: &tfprotov6.Schema{
-				Version: 1,
-				Block: &tfprotov6.SchemaBlock{
-					Attributes: []*tfprotov6.SchemaAttribute{
-						{
-							Name:     "string",
-							Type:     tftypes.String,
-							Required: true,
-						},
-					},
-					Description:     "a test resource",
-					DescriptionKind: tfprotov6.StringKindPlain,
-				},
-			},
-		},
-		"deprecated": {
-			input: schema.Schema{
-				Version: 1,
-				Attributes: map[string]schema.Attribute{
-					"string": {
-						Type:     types.StringType,
-						Required: true,
-					},
-				},
-				DeprecationMessage: "deprecated, use other_resource instead",
-			},
-			expected: &tfprotov6.Schema{
-				Version: 1,
-				Block: &tfprotov6.SchemaBlock{
-					Attributes: []*tfprotov6.SchemaAttribute{
-						{
-							Name:     "string",
-							Type:     tftypes.String,
-							Required: true,
-						},
-					},
-					Deprecated: true,
-				},
-			},
-		},
-	}
-
-	for name, tc := range tests {
-		name, tc := name, tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			got, err := Schema(context.Background(), tc.input)
-			if err != nil {
-				if tc.expectedErr == "" {
-					t.Errorf("Unexpected error: %s", err)
-					return
-				}
-				if err.Error() != tc.expectedErr {
-					t.Errorf("Expected error to be %q, got %q", tc.expectedErr, err.Error())
-					return
-				}
-				// got expected error
-				return
-			}
-			if err == nil && tc.expectedErr != "" {
-				t.Errorf("Expected error to be %q, got nil", tc.expectedErr)
-				return
-			}
-			if diff := cmp.Diff(got, tc.expected); diff != "" {
-				t.Errorf("Unexpected diff (+wanted, -got): %s", diff)
-				return
-			}
-		})
-	}
-}
-
-func TestAttribute(t *testing.T) {
+func TestAttributeTfprotov6SchemaAttribute(t *testing.T) {
 	t.Parallel()
 
 	type testCase struct {
 		name        string
-		attr        schema.Attribute
+		attr        Attribute
 		path        *tftypes.AttributePath
 		expected    *tfprotov6.SchemaAttribute
 		expectedErr string
@@ -466,7 +25,7 @@ func TestAttribute(t *testing.T) {
 	tests := map[string]testCase{
 		"deprecated": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:               types.StringType,
 				Optional:           true,
 				DeprecationMessage: "deprecated, use new_string instead",
@@ -481,7 +40,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"description-plain": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:        types.StringType,
 				Optional:    true,
 				Description: "A string attribute",
@@ -497,7 +56,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"description-markdown": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:                types.StringType,
 				Optional:            true,
 				MarkdownDescription: "A string attribute",
@@ -513,7 +72,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"description-both": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:                types.StringType,
 				Optional:            true,
 				Description:         "A string attribute",
@@ -530,7 +89,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-string": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.StringType,
 				Optional: true,
 			},
@@ -543,7 +102,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-bool": {
 			name: "bool",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.BoolType,
 				Optional: true,
 			},
@@ -556,7 +115,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-number": {
 			name: "number",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.NumberType,
 				Optional: true,
 			},
@@ -569,7 +128,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-list": {
 			name: "list",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.ListType{ElemType: types.NumberType},
 				Optional: true,
 			},
@@ -582,7 +141,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-map": {
 			name: "map",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.MapType{ElemType: types.StringType},
 				Optional: true,
 			},
@@ -595,7 +154,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-object": {
 			name: "object",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type: types.ObjectType{AttrTypes: map[string]attr.Type{
 					"foo": types.StringType,
 					"bar": types.NumberType,
@@ -618,7 +177,7 @@ func TestAttribute(t *testing.T) {
 		// TODO: add tuple attribute when we support it
 		"required": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.StringType,
 				Required: true,
 			},
@@ -631,7 +190,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"optional": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.StringType,
 				Optional: true,
 			},
@@ -644,7 +203,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"computed": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.StringType,
 				Computed: true,
 			},
@@ -657,7 +216,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"optional-computed": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:     types.StringType,
 				Computed: true,
 				Optional: true,
@@ -672,7 +231,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"sensitive": {
 			name: "string",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type:      types.StringType,
 				Optional:  true,
 				Sensitive: true,
@@ -687,8 +246,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-single": {
 			name: "single_nested",
-			attr: schema.Attribute{
-				Attributes: schema.SingleNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: SingleNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -725,8 +284,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-list": {
 			name: "list_nested",
-			attr: schema.Attribute{
-				Attributes: schema.ListNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: ListNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -736,7 +295,7 @@ func TestAttribute(t *testing.T) {
 						Computed:  true,
 						Sensitive: true,
 					},
-				}, schema.ListNestedAttributesOptions{}),
+				}, ListNestedAttributesOptions{}),
 				Optional: true,
 			},
 			path: tftypes.NewAttributePath(),
@@ -763,8 +322,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-list-min": {
 			name: "list_nested",
-			attr: schema.Attribute{
-				Attributes: schema.ListNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: ListNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -774,7 +333,7 @@ func TestAttribute(t *testing.T) {
 						Computed:  true,
 						Sensitive: true,
 					},
-				}, schema.ListNestedAttributesOptions{
+				}, ListNestedAttributesOptions{
 					MinItems: 1,
 				}),
 				Optional: true,
@@ -804,8 +363,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-list-max": {
 			name: "list_nested",
-			attr: schema.Attribute{
-				Attributes: schema.ListNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: ListNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -815,7 +374,7 @@ func TestAttribute(t *testing.T) {
 						Computed:  true,
 						Sensitive: true,
 					},
-				}, schema.ListNestedAttributesOptions{
+				}, ListNestedAttributesOptions{
 					MaxItems: 1,
 				}),
 				Optional: true,
@@ -845,8 +404,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-list-minmax": {
 			name: "list_nested",
-			attr: schema.Attribute{
-				Attributes: schema.ListNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: ListNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -856,7 +415,7 @@ func TestAttribute(t *testing.T) {
 						Computed:  true,
 						Sensitive: true,
 					},
-				}, schema.ListNestedAttributesOptions{
+				}, ListNestedAttributesOptions{
 					MinItems: 1,
 					MaxItems: 10,
 				}),
@@ -888,8 +447,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-set": {
 			name: "set_nested",
-			attr: schema.Attribute{
-				Attributes: schema.SetNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: SetNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -899,7 +458,7 @@ func TestAttribute(t *testing.T) {
 						Computed:  true,
 						Sensitive: true,
 					},
-				}, schema.SetNestedAttributesOptions{}),
+				}, SetNestedAttributesOptions{}),
 				Optional: true,
 			},
 			path: tftypes.NewAttributePath(),
@@ -926,8 +485,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-set-min": {
 			name: "set_nested",
-			attr: schema.Attribute{
-				Attributes: schema.SetNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: SetNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -937,7 +496,7 @@ func TestAttribute(t *testing.T) {
 						Computed:  true,
 						Sensitive: true,
 					},
-				}, schema.SetNestedAttributesOptions{
+				}, SetNestedAttributesOptions{
 					MinItems: 1,
 				}),
 				Optional: true,
@@ -967,8 +526,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-set-max": {
 			name: "set_nested",
-			attr: schema.Attribute{
-				Attributes: schema.SetNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: SetNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -978,7 +537,7 @@ func TestAttribute(t *testing.T) {
 						Computed:  true,
 						Sensitive: true,
 					},
-				}, schema.SetNestedAttributesOptions{
+				}, SetNestedAttributesOptions{
 					MaxItems: 1,
 				}),
 				Optional: true,
@@ -1008,8 +567,8 @@ func TestAttribute(t *testing.T) {
 		},
 		"nested-attr-set-minmax": {
 			name: "set_nested",
-			attr: schema.Attribute{
-				Attributes: schema.SetNestedAttributes(map[string]schema.Attribute{
+			attr: Attribute{
+				Attributes: SetNestedAttributes(map[string]Attribute{
 					"string": {
 						Type:     types.StringType,
 						Optional: true,
@@ -1019,7 +578,7 @@ func TestAttribute(t *testing.T) {
 						Computed:  true,
 						Sensitive: true,
 					},
-				}, schema.SetNestedAttributesOptions{
+				}, SetNestedAttributesOptions{
 					MinItems: 1,
 					MaxItems: 10,
 				}),
@@ -1051,9 +610,9 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-and-nested-attr-set": {
 			name: "whoops",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Type: types.StringType,
-				Attributes: schema.SingleNestedAttributes(map[string]schema.Attribute{
+				Attributes: SingleNestedAttributes(map[string]Attribute{
 					"testing": {
 						Type:     types.StringType,
 						Optional: true,
@@ -1066,7 +625,7 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-and-nested-attr-unset": {
 			name: "whoops",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Optional: true,
 			},
 			path:        tftypes.NewAttributePath(),
@@ -1074,9 +633,9 @@ func TestAttribute(t *testing.T) {
 		},
 		"attr-and-nested-attr-empty": {
 			name: "whoops",
-			attr: schema.Attribute{
+			attr: Attribute{
 				Optional:   true,
-				Attributes: schema.SingleNestedAttributes(map[string]schema.Attribute{}),
+				Attributes: SingleNestedAttributes(map[string]Attribute{}),
 			},
 			path:        tftypes.NewAttributePath(),
 			expectedErr: "must have Attributes or Type set",
@@ -1088,7 +647,7 @@ func TestAttribute(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			got, err := Attribute(context.Background(), tc.name, tc.attr, tc.path)
+			got, err := tc.attr.tfprotov6SchemaAttribute(context.Background(), tc.name, tc.path)
 			if err != nil {
 				if tc.expectedErr == "" {
 					t.Errorf("Unexpected error: %s", err)
